Match utf32 char code parameter type

+++ cocos/2d/CCFontFreeType.h
@@ -93,7 +93,7 @@ public:
 
     int* getHorizontalKerningForTextUTF32(const std::u32string& text, int& outNumLetters) const override;
 
-    unsigned char* getGlyphBitmap(uint64_t theChar, int32_t& outWidth, int32_t& outHeight, Rect& outRect, int& xAdvance);
+    unsigned char* getGlyphBitmap(uint32_t theChar, int32_t& outWidth, int32_t& outHeight, Rect& outRect, int& xAdvance);
 
     int getFontAscender() const;
     const char* getFontFamily() const;
